ctr: move unpack to snapshot command
